Cleanup
RulesHowToPlay is now much easier to read + bluff- and actionchoices buttons have been moved into components, so that rules and the game itself share components. Will do a similar thing with health and turn information not only to explain the rules but also so that all game elements from the local version is now available everywhere, so that the online version can use the same components.