Responsive rejoining
Any user who left without clicking 'leave game', e.g., accidentally closed the browser window or otherwise got the socket to disconnect, they now have 2min to reconnect. I also made it so that when the server resets, the clients who are connected will be informed and possibly redirected when the sockets rejoin.